Split grammar into main grammar and expression grammar
